requirements.txt
streamlit
python-dotenv
psycopg2-binary
pgvector
pandas
sentence-transformers
langchain
google-generativeai
